refactor: introduce `plaiceholder` package
Migrate all functions to a single, consumable package
