+++ docs/pages/plugins.mdx
@@ -57,7 +57,7 @@ We'll then apply the string to an `<img>` element (hidden from screen-readers) a
    import * as React from "react";
    import Image from "next/image";
    import { getImage } from "@plaiceholder/next";
-   import { getBase64 } from "@plaiceholder/base64";
+   import { getBase64 } from "plaiceholder";
 
    export const getStaticProps: GetStaticProps = async () => {
      const imgSrc = "/keila-joa.jpg";
