+++ examples/next/package.json
@@ -14,14 +14,12 @@
   },
   "dependencies": {
     "@joebell/cx": "0.1.0",
-    "@plaiceholder/base64": "^1.0.0",
-    "@plaiceholder/blurhash": "^1.0.0",
-    "@plaiceholder/css": "^1.0.0",
-    "@plaiceholder/svg": "^1.0.0",
+    "@plaiceholder/next": "^1.0.0",
     "@plaiceholder/tailwindcss": "^1.0.0",
     "@tailwindcss/aspect-ratio": "0.2.0",
     "glob": "7.1.6",
     "next": "10.2.0",
+    "plaiceholder": "^1.0.0",
     "react": "17.0.2",
     "react-blurhash": "0.1.3",
     "react-dom": "17.0.2",
