+++ examples/next/src/pages/svg/multiple.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { InferGetStaticPropsType } from "next";
 import Image from "next/image";
 import { getImage } from "@plaiceholder/next";
-import { getPixelsSVG } from "@plaiceholder/svg";
+import { getPixelsSVG } from "plaiceholder";
 import { config } from "@/config";
 import { Layout } from "@/components/layout";
 import { getAllPublicImagePaths } from "@/lib/images";
