+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "packages": ["docs", "examples/*", "packages/*", "plugins/*"],
+  "packages": ["docs", "examples/*", "packages/*"],
   "version": "1.0.0",
   "npmClient": "yarn",
   "useWorkspaces": true,
