--- packages/css/CHANGELOG.md
@@ -1,10 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [1.0.0](https://github.com/joe-bell/plaiceholder/compare/v0.2.0...v1.0.0) (2020-11-25)
-
-### Bug Fixes
-
-- move rgba to CSS ([1ec2cfb](https://github.com/joe-bell/plaiceholder/commit/1ec2cfbe6e47b78d78ecb323db52819a1fba3574))
