+++ packages/plaiceholder/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@plaiceholder/blurhash",
+  "name": "plaiceholder",
   "version": "1.0.0",
-  "description": "Roll-you-own Blurhash placeholders.",
+  "description": "Beautiful image placeholders, without the hassle.",
   "author": "Joe Bell <joe@joebell.co.uk>",
   "license": "Apache-2.0",
   "funding": "https://plaiceholder.co",
@@ -26,17 +26,20 @@
     "lqip",
     "placeholder",
     "plaiceholder",
-    "blurhash"
+    "base64",
+    "blurhash",
+    "css",
+    "svg"
   ],
   "bugs": {
     "url": "https://github.com/joe-bell/plaiceholder/issues"
   },
   "homepage": "https://github.com/joe-bell/plaiceholder#readme",
-  "dependencies": {
-    "@plaiceholder/core": "^1.0.0",
-    "blurhash": "1.1.3"
-  },
   "peerDependencies": {
     "sharp": ">= 0.26.2"
+  },
+  "dependencies": {
+    "@plaiceholder/core": "1.0.0",
+    "blurhash": "1.1.3"
   }
 }
