+++ packages/plaiceholder/src/svg.ts
