--- packages/svg/CHANGELOG.md
@@ -1,8 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# [1.0.0](https://github.com/joe-bell/plaiceholder/compare/v0.2.0...v1.0.0) (2020-11-25)
-
-**Note:** Version bump only for package @plaiceholder/svg
