--- packages/svg/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@plaiceholder/svg",
-  "version": "1.0.0",
-  "description": "Roll-you-own SVG placeholders.",
-  "author": "Joe Bell <joe@joebell.co.uk>",
-  "license": "Apache-2.0",
-  "funding": "https://plaiceholder.co",
-  "source": "src/index.ts",
-  "main": "dist/index.js",
-  "module": "dist/index.esm.js",
-  "typings": "dist/index.d.ts",
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "microbundle --no-compress --tsconfig tsconfig.json --external sharp",
-    "dev": "yarn build -- watch",
-    "clean": "rimraf {dist,.rts2*}"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/joe-bell/plaiceholder.git"
-  },
-  "keywords": [
-    "lqip",
-    "placeholder",
-    "plaiceholder",
-    "svg"
-  ],
-  "bugs": {
-    "url": "https://github.com/joe-bell/plaiceholder/issues"
-  },
-  "homepage": "https://github.com/joe-bell/plaiceholder#readme",
-  "dependencies": {
-    "@plaiceholder/core": "^1.0.0"
-  },
-  "peerDependencies": {
-    "sharp": ">= 0.26.2"
-  }
-}
