--- packages/svg/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "exclude": ["node_modules"],
-  "include": ["src/**/*.ts"]
-}
