+++ packages/tailwindcss/src/config.ts
